docs(site): Upgrade to React Router v4
