+++ docs/src/components/App/App.js
@@ -1,20 +1,28 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-
+import { Route } from 'react-router';
 import { StyleGuideProvider } from 'seek-style-guide/react';
 import Header from './Header/Header';
 import Footer from './Footer/Footer';
+import Home from 'Home/Home';
+import PageLayout from 'PageLayout/PageLayout';
+import Typography from 'Typography/Typography';
+import Playground from 'Playground/Playground';
+import Demo from 'Demo/Demo';
+import demoSpecs from '../../demoSpecs';
 
-export default function App({ children }) {
-  return (
-    <StyleGuideProvider fullScreen={true} title="SEEK Style Guide">
-      <Header />
-      {children}
-      <Footer />
-    </StyleGuideProvider>
-  );
-}
-
-App.propTypes = {
-  children: PropTypes.node
-};
+export default () => (
+  <StyleGuideProvider fullScreen={true} title="SEEK Style Guide">
+    <Header />
+    <Route path="/" exact component={Home} />
+    <Route path="/typography" component={Typography} />
+    <Route path="/page-layout" component={PageLayout} />
+    <Route path="/playground" component={Playground} />
+    {
+      demoSpecs.map(demoSpec => {
+        const DemoRoute = () => <Demo spec={demoSpec} />;
+        return <Route key={demoSpec.title} path={demoSpec.route} component={DemoRoute} />;
+      })
+    }
+    <Footer />
+  </StyleGuideProvider>
+);
