+++ docs/src/components/App/Header/Header.js
@@ -3,7 +3,7 @@ import styles from './Header.less';
 import React, { Component } from 'react';
 import classnames from 'classnames';
 
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import { PageBlock, Card, Section, Text, ScreenReaderOnly } from 'seek-style-guide/react';
 import Logo from './Logo/Logo';
 import demoSpecs from '../../../demoSpecs';
