+++ package.json
@@ -102,7 +102,6 @@
     "gh-pages": "^0.11.0",
     "glob": "^7.1.1",
     "hex-rgb": "^1.0.0",
-    "history": "^3.3.0",
     "html-webpack-plugin": "^2.28.0",
     "husky": "^0.12.0",
     "identity-obj-proxy": "^3.0.0",
@@ -121,7 +120,8 @@
     "react-element-to-jsx-string": "^12.0.0",
     "react-helmet": "^5.2.0",
     "react-hot-loader": "^3.0.0-beta.2",
-    "react-router": "^3.2.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-shallow-testutils": "^3.0.0",
     "react-test-renderer": "^16.0.0",
     "rewrite-files": "^1.0.1",
