+++ react/Header/Header.demo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import { Header } from 'seek-style-guide/react';
 import LogoRainbow from '../LogoRainbow/LogoRainbow';
